+++ package.json
@@ -22,7 +22,7 @@
   },
   "dependencies": {
     "@mintbase-js/react": "0.5.4-beta.0",
-    "near-ca": "^0.3.0",
+    "near-ca": "^0.3.1",
     "next": "14.0.1",
     "react": "^18",
     "react-dom": "^18",
